# frontend/openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_sound": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_sound",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "SoundKind": {
        "description": "The audio-source discriminator. Stored as checked text, not a native enum.",
        "enum": [
          "file",
          "youtube"
        ],
        "title": "SoundKind",
        "type": "string"
      },
      "SoundRead": {
        "description": "A single library entry as returned by the API.",
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_errored": {
            "title": "Is Errored",
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/SoundKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "youtube_video_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Video Id"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "duration_seconds",
          "is_errored",
          "error_detail",
          "youtube_video_id",
          "content_type",
          "created_at"
        ],
        "title": "SoundRead",
        "type": "object"
      },
      "TagRead": {
        "description": "A single tag as returned by the API.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "TagRead",
        "type": "object"
      },
      "TagWrite": {
        "description": "Body of `POST /api/tags` and `PATCH /api/tags/{id}` \u2014 both just set `name`.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TagWrite",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "YoutubeAddRequest": {
        "description": "Body of `POST /api/sounds/youtube`.",
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "YoutubeAddRequest",
        "type": "object"
      },
      "YoutubeSoundRead": {
        "description": "`POST /api/sounds/youtube`'s response: a `SoundRead` plus the add-time\nembeddability warning (ADR-0005). Non-null only when the oEmbed heuristic\nreturned a 401 (\"owner may have disabled embedding\") \u2014 the sound is still\ncreated, but the client can use this to show an \"Add anyway\"-style notice.",
        "properties": {
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "embed_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Embed Warning"
          },
          "error_detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Detail"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_errored": {
            "title": "Is Errored",
            "type": "boolean"
          },
          "kind": {
            "$ref": "#/components/schemas/SoundKind"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "youtube_video_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Youtube Video Id"
          }
        },
        "required": [
          "id",
          "name",
          "kind",
          "duration_seconds",
          "is_errored",
          "error_detail",
          "youtube_video_id",
          "content_type",
          "created_at"
        ],
        "title": "YoutubeSoundRead",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "ElectroBard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/sounds": {
      "get": {
        "description": "List the current user's sounds, A-Z by name.\n\n`?tag=` filters to sounds carrying that tag id; `?q=` is an optional\ncase-insensitive substring search on `name`. Both are scoped to the current user.",
        "operationId": "list_sounds",
        "parameters": [
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SoundRead"
                  },
                  "title": "Response List Sounds",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sounds",
        "tags": [
          "sounds"
        ]
      }
    },
    "/api/sounds/upload": {
      "post": {
        "description": "Multipart upload -> one `file` Sound (api-contract \"Upload ingestion\", #22).\n\nFormat is validated by the filename's extension against a fixed allowlist \u2014 the\nclient's declared Content-Type is never trusted. The upload is buffered fully\n(no streaming, Q6) and handed to the bytes-only storage seam. The Sound row is\ncreated and flushed first to mint its UUID (so `storage_key` can embed it), and\nthe whole handler runs in `get_db`'s single request transaction: if `storage.save`\nraises, the exception propagates and that transaction rolls back, leaving no\norphan row.",
        "operationId": "upload_sound",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_sound"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoundRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Sound",
        "tags": [
          "sounds"
        ]
      }
    },
    "/api/sounds/youtube": {
      "post": {
        "description": "`{ url }` -> one `youtube` Sound, keyless (ADR-0005, api-contract \"Sounds\").\n\nThe video ID is extracted by structural URL parse (no API key); metadata comes\nfrom YouTube's own oEmbed endpoint, which carries a title but never a duration \u2014\n`duration_seconds` stays null (client `getDuration()` backfill is post-M1).\nAdd-time embeddability is a *heuristic* on the oEmbed status, not the final\nverdict (that's the client IFrame `onError` at playback, M3/#25): 200 accepts,\n401 still accepts but flags `embed_warning`, and 400/404 reject as unusable.",
        "operationId": "add_youtube_sound",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/YoutubeAddRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/YoutubeSoundRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Youtube Sound",
        "tags": [
          "sounds"
        ]
      }
    },
    "/api/sounds/{sound_id}": {
      "get": {
        "description": "Fetch one sound. 404 if missing or owned by another user.",
        "operationId": "get_sound",
        "parameters": [
          {
            "in": "path",
            "name": "sound_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Sound Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoundRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Sound",
        "tags": [
          "sounds"
        ]
      }
    },
    "/api/tags": {
      "get": {
        "description": "List the current user's tags, A-Z by name.",
        "operationId": "list_tags",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TagRead"
                  },
                  "title": "Response List Tags",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Tags",
        "tags": [
          "tags"
        ]
      },
      "post": {
        "description": "Create a tag. 409 if the user already has one with this name.",
        "operationId": "create_tag",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Tag",
        "tags": [
          "tags"
        ]
      }
    },
    "/api/tags/{tag_id}": {
      "delete": {
        "description": "Delete a tag. Drops it from every Sound (`sound_tags` cascades at the DB level\nvia `ON DELETE CASCADE`); a Sound (or, later, a Set) may end up with no tags and\njust stays that way \u2014 post-M1 concern. 404 if missing/wrong tenant.",
        "operationId": "delete_tag",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tag Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Tag",
        "tags": [
          "tags"
        ]
      },
      "patch": {
        "description": "Rename a tag. 404 if missing/wrong tenant; 409 on a name collision.",
        "operationId": "rename_tag",
        "parameters": [
          {
            "in": "path",
            "name": "tag_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Tag Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagWrite"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Tag",
        "tags": [
          "tags"
        ]
      }
    },
    "/health": {
      "get": {
        "description": "Liveness: the process is up and serving.",
        "operationId": "health",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/health/db": {
      "get": {
        "description": "Readiness: the database is reachable.",
        "operationId": "health_db",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Health Db",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health Db"
      }
    }
  }
}